requirements-prod.txt
# Production-ready requirements
# Core ML/DL
tensorflow==2.15.0
numpy==1.24.3
opencv-python-headless==4.8.1.78
Pillow==10.1.0
scikit-learn==1.3.2
scipy==1.11.4
matplotlib==3.8.2
seaborn==0.13.0
pandas==2.1.4

# Web API
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.2

# Gradio demo
gradio==4.44.0

# MLOps & Monitoring
mlflow==2.9.2
wandb==0.16.1
pyyaml==6.0.1

# Testing
pytest==7.4.3
pytest-cov==4.1.0
pytest-asyncio==0.21.1
httpx==0.25.2

# Code quality
black==23.12.0
flake8==6.1.0
mypy==1.7.1
pylint==3.0.3

# Utilities
python-dotenv==1.0.0
tqdm==4.66.1
pydicom==2.4.4  # For DICOM support
